refactor: (#191) seat & study-room port 합치기
