--- dms-application/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/CommandSeatPort.kt
@@ -1,9 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-import team.aliens.dms.domain.studyroom.model.Seat
-
-interface CommandSeatPort {
-
-    fun saveSeat(seat: Seat): Seat
-
-}
