+++ dms-application/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/QueryStudyRoomPort.kt
@@ -1,10 +1,13 @@
 package team.aliens.dms.domain.studyroom.spi
 
 import java.util.UUID
+import team.aliens.dms.domain.studyroom.model.Seat
 import team.aliens.dms.domain.studyroom.model.StudyRoom
 
 interface QueryStudyRoomPort {
 
     fun queryStudyRoomById(studyRoomId: UUID): StudyRoom?
+
+    fun querySeatById(seatId: UUID): Seat?
 
 }
