--- dms-application/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/SeatPort.kt
@@ -1,6 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-interface SeatPort :
-    QuerySeatPort,
-    CommandSeatPort {
-}
