+++ dms-application/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/StudyRoomPort.kt
@@ -1,5 +1,6 @@
 package team.aliens.dms.domain.studyroom.spi
 
 interface StudyRoomPort:
-    QueryStudyRoomPort {
+    QueryStudyRoomPort,
+    CommandStudyRoomPort {
 }
