+++ dms-application/src/main/kotlin/team/aliens/dms/domain/studyroom/usecase/ApplySeatUseCase.kt
@@ -6,9 +6,7 @@ import team.aliens.dms.domain.school.exception.SchoolMismatchException
 import team.aliens.dms.domain.studyroom.exception.SeatAlreadyAppliedException
 import team.aliens.dms.domain.studyroom.exception.SeatNotFoundException
 import team.aliens.dms.domain.studyroom.exception.StudyRoomNotFoundException
-import team.aliens.dms.domain.studyroom.model.Seat
-import team.aliens.dms.domain.studyroom.spi.CommandSeatPort
-import team.aliens.dms.domain.studyroom.spi.QuerySeatPort
+import team.aliens.dms.domain.studyroom.spi.CommandStudyRoomPort
 import team.aliens.dms.domain.studyroom.spi.QueryStudyRoomPort
 import team.aliens.dms.domain.studyroom.spi.StudyRoomQueryUserPort
 import team.aliens.dms.domain.studyroom.spi.StudyRoomSecurityPort
@@ -18,16 +16,15 @@ import team.aliens.dms.domain.user.exception.UserNotFoundException
 class ApplySeatUseCase(
     private val securityPort: StudyRoomSecurityPort,
     private val queryUserPort: StudyRoomQueryUserPort,
-    private val querySeatPort: QuerySeatPort,
     private val queryStudyRoomPort: QueryStudyRoomPort,
-    private val commandSeatPort: CommandSeatPort
+    private val commandStudyRoomPort: CommandStudyRoomPort
 ) {
 
     fun execute(seatId: UUID) {
         val currentUserId = securityPort.getCurrentUserId()
         val user = queryUserPort.queryUserById(currentUserId) ?: throw UserNotFoundException
 
-        val seat = querySeatPort.querySeatById(seatId) ?: throw SeatNotFoundException
+        val seat = queryStudyRoomPort.querySeatById(seatId) ?: throw SeatNotFoundException
         val studyRoom = queryStudyRoomPort.queryStudyRoomById(seat.studyRoomId) ?: throw StudyRoomNotFoundException
 
         if (user.schoolId != studyRoom.schoolId) {
@@ -39,6 +36,6 @@ class ApplySeatUseCase(
         } ?: run {
             seat.copy(studentId = currentUserId)
         }
-        commandSeatPort.saveSeat(saveSeat)
+        commandStudyRoomPort.saveSeat(saveSeat)
     }
 }
