--- dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/studyroom/SeatPersistenceAdapter.kt
@@ -1,26 +1,0 @@
-package team.aliens.dms.persistence.studyroom
-
-import java.util.UUID
-import org.springframework.data.repository.findByIdOrNull
-import org.springframework.stereotype.Component
-import team.aliens.dms.domain.studyroom.model.Seat
-import team.aliens.dms.domain.studyroom.spi.SeatPort
-import team.aliens.dms.persistence.studyroom.mapper.SeatMapper
-import team.aliens.dms.persistence.studyroom.repository.SeatJpaRepository
-
-@Component
-class SeatPersistenceAdapter(
-    private val seatMapper: SeatMapper,
-    private val seatRepository: SeatJpaRepository
-) : SeatPort {
-
-    override fun querySeatById(seatId: UUID) = seatMapper.toDomain(
-        seatRepository.findByIdOrNull(seatId)
-    )
-
-    override fun saveSeat(seat: Seat) = seatMapper.toDomain(
-        seatRepository.save(
-            seatMapper.toEntity(seat)
-        )
-    )!!
-}
