+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/studyroom/StudyRoomPersistenceAdapter.kt
@@ -3,17 +3,32 @@ package team.aliens.dms.persistence.studyroom
 import java.util.UUID
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
+import team.aliens.dms.domain.studyroom.model.Seat
 import team.aliens.dms.domain.studyroom.spi.StudyRoomPort
+import team.aliens.dms.persistence.studyroom.mapper.SeatMapper
 import team.aliens.dms.persistence.studyroom.mapper.StudyRoomMapper
+import team.aliens.dms.persistence.studyroom.repository.SeatJpaRepository
 import team.aliens.dms.persistence.studyroom.repository.StudyRoomJpaRepository
 
 @Component
 class StudyRoomPersistenceAdapter(
     private val studyRoomMapper: StudyRoomMapper,
-    private val studyRoomRepository: StudyRoomJpaRepository
+    private val seatMapper: SeatMapper,
+    private val studyRoomRepository: StudyRoomJpaRepository,
+    private val seatRepository: SeatJpaRepository
 ) : StudyRoomPort {
 
     override fun queryStudyRoomById(studyRoomId: UUID) = studyRoomMapper.toDomain(
         studyRoomRepository.findByIdOrNull(studyRoomId)
     )
+
+    override fun querySeatById(seatId: UUID) = seatMapper.toDomain(
+        seatRepository.findByIdOrNull(seatId)
+    )
+
+    override fun saveSeat(seat: Seat) = seatMapper.toDomain(
+        seatRepository.save(
+            seatMapper.toEntity(seat)
+        )
+    )!!
 }
